requirements.txt
discord.py>=2.3.2
openai>=1.30.0
selenium>=4.20.0
webdriver-manager>=3.9.1
python-dotenv>=1.0.1
selenium-stealth
